+++ src/utils/helper.ts
@@ -1,3 +1,6 @@
+import { ShellProfile } from "../types/MessageProtocol";
+import { SHELL_PROFILES } from "./constants";
+
 /**
  * Generate a random ID for blocks or other entities
  */
@@ -14,4 +17,30 @@ export function getNonce() {
     text += possible.charAt(Math.floor(Math.random() * possible.length));
   }
   return text;
+}
+
+/**
+ * Default Shell
+ */
+export function getDefaultShell(): ShellProfile {
+  if (process.platform === "win32") {
+    // prefer pwsh > powershell > cmd
+    return (
+      SHELL_PROFILES.find((s) => s.id === "pwsh") ||
+      SHELL_PROFILES.find((s) => s.id === "powershell") ||
+      SHELL_PROFILES.find((s) => s.id === "cmd")!
+    );
+  }
+
+  const envShell = process.env.SHELL;
+
+  if (envShell) {
+    const match = SHELL_PROFILES.find((s) => envShell.includes(s.command));
+    if (match) {
+      return match;
+    }
+  }
+
+  // fallback to bash
+  return SHELL_PROFILES.find((s) => s.id === "bash")!;
 }